refactor: split benchmark command into perf and quality modules
Pure restructure ahead of the Phase 0 harness work: the 633-line
benchmark.rs becomes benchmark/{mod,perf,quality}.rs with dispatch in
mod.rs. Drops the dead description/rationale fields from the benchmark
file structs (serde ignores unknown JSON fields). No behavior change.
